change double in time Meal.java to int, add 'minut' in time table in meal.html

+++ src/main/java/com/agiklo/HeathProject/model/Meal.java
@@ -20,5 +20,5 @@ public class Meal {
     private Long id;
     private String name;
     private String description;
-    private Double time;
+    private Integer time;
 }
