+++ src/main/java/com/agiklo/HeathProject/repository/MealRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MealRepository extends JpaRepository<Meal, Long> {
     Meal findByName(String name);
-    Meal findByTime(double time);
+    Meal findByTime(int time);
 }
